Add model to simulate task relation
There are now two ways tasks can be related to one another.

Through the `task_id` on the `tasks` table:
This is hierarchial. A task can only be a child of one other task

Through `task_dependencies` table:
This way, a task can be related to multiple tasks. Here, none is above another.
A validation prevents bidirectional dependency.